+++ R/kalfil.R
@@ -192,7 +192,7 @@ kalfil_update_params <- function(alpha_j, fwd, setup) {
   phase <- numeric(k)
   for (j in seq_len(k)) {
     s <- alpha_j[2 * j + 1, 1:n]
-    pgr <- peri(s, plot = FALSE)
+    pgr <- peri(s)
     loc <- which.min(abs(pgr$c - f[j]))
     amp[j] <- pgr$amp[loc]
     phase[j] <- pgr$phase[loc]